Use a static list if there are no custom classes
Supersedes #43

This resolves a crash multiple customers have been seeing. It is unclear
to me what the root cause is, because I couldn't reproduce it myself and
it is reportedly happening randomly. The crash will still happen in case
the user actually does use custom classes, presumably, but it is at
least a mitigation for the issue.