Fix a potential race condition regarding localisation
Prior to this PR, the localisation flag existed on the client object and resources were querying it. This leaves up the potential for subsequent operations to change the flag before all resources which depend on it have been fully processed. This change injects the localisation flag at construction time instead, so that it is immutable for a given resource.